Use more `Stream`s for `DepartmentFloor`

+++ java/solutions/src/main/java/com/gusmurphy/fun/aoc/year2025/day4/DepartmentFloorSolver.java
@@ -1,7 +1,7 @@
 package com.gusmurphy.fun.aoc.year2025.day4;
 
 public class DepartmentFloorSolver {
-    public static int accessibleRollCountFromFile(String absolutePath) {
+    public static long accessibleRollCountFromFile(String absolutePath) {
         var floor = FloorGridParser.parseFile(absolutePath);
         return floor.accessibleRollCount();
     }
